Crate libobs_sources

A library for creating OBS sources without having to figure out what properties are used by sources. Example usage (for window capture only on windows):

use libobs_window_helper::WindowSearchMode;
use libobs_wrapper::{context::ObsContext, sources::ObsSourceBuilder, utils::{OutputInfo, StartupInfo}};
use libobs_sources::windows::WindowCaptureSourceBuilder;

let output = context.output(output_info).unwrap();

// Do other initialization for video encoders, audio encoders, etc.

let windows = WindowCaptureSourceBuilder::get_windows(WindowSearchMode::IncludeMinimized).unwrap();
let example_window = windows.get(0).unwrap();

WindowCaptureSourceBuilder::new("Test Window Capture")
.set_window(example_window)
// Obs Output is created from `ObsContext`
.add_to_output(output, 0)
.unwrap();

Modules

encoders

output

windows

Traits

ObsObjectUpdater

ObsSourceBuilder