libobs_sources

Module windows

Structs

GameCaptureSourceBuilder

GameCaptureSourceUpdater

Used to update the source this updater was created from. For more details look at docs for the corresponding builder.

MonitorCaptureSourceBuilder

Provides a easy to use builder for the monitor capture source.

MonitorCaptureSourceUpdater

Used to update the source this updater was created from. For more details look at docs for the corresponding builder.

WindowCaptureSourceBuilder

Provides a easy to use builder for the window capture source.

WindowCaptureSourceUpdater

Used to update the source this updater was created from. For more details look at docs for the corresponding builder.

Enums

ObsDisplayCaptureMethod

Describes the capture method of the monitor capture source. Used in MonitorCaptureSourceBuilder

ObsGameCaptureMode

Describes the capture mode of the game capture source.

ObsGameCaptureRgbaSpace

ObsHookRate

Describes the capture method of the window capture source. Used in WindowCaptureSourceBuilder

ObsWindowCaptureMethod

Describes the capture method of the window capture source. Used in WindowCaptureSourceBuilder

ObsWindowPriority

Describes the priority of the window capture source. Used in WindowCaptureSourceBuilder